feat: SolutionReviewsResponse 추가
Related to : #9
