+++ backend/src/main/java/com/prograngers/backend/controller/ProblemController.java
@@ -1,15 +1,13 @@
 package com.prograngers.backend.controller;
 
 
-import com.prograngers.backend.dto.ProblemListResponse;
-import com.prograngers.backend.entity.Problem;
+import com.prograngers.backend.dto.problem.ProblemListResponse;
 import com.prograngers.backend.service.ProblemService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
