+++ backend/src/main/java/com/prograngers/backend/dto/review/Line.java
@@ -1,0 +1,4 @@
+package com.prograngers.backend.dto.review;
+
+public class Line {
+}
