+++ backend/src/main/java/com/prograngers/backend/dto/solution/SolutionDetailResponseComment.java
@@ -1,4 +1,4 @@
-package com.prograngers.backend.dto;
+package com.prograngers.backend.dto.solution;
 
 
 import lombok.AllArgsConstructor;
