+++ backend/src/test/java/com/prograngers/backend/service/CommentServiceTest.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.service;
 
-import com.prograngers.backend.dto.CommentPatchRequest;
+import com.prograngers.backend.dto.comment.CommentPatchRequest;
 import com.prograngers.backend.entity.Comment;
 import com.prograngers.backend.entity.Solution;
 import com.prograngers.backend.exception.notfound.CommentNotFoundException;
