+++ backend/src/test/java/com/prograngers/backend/service/SolutionServiceTest.java
@@ -1,7 +1,7 @@
 package com.prograngers.backend.service;
 
-import com.prograngers.backend.dto.ScarpSolutionRequest;
-import com.prograngers.backend.dto.SolutionPatchRequest;
+import com.prograngers.backend.dto.solution.ScarpSolutionRequest;
+import com.prograngers.backend.dto.solution.SolutionPatchRequest;
 import com.prograngers.backend.entity.Member;
 import com.prograngers.backend.entity.Problem;
 import com.prograngers.backend.entity.Solution;
